Fixed int instead of float

+++ include/tuple.hpp
@@ -17,7 +17,7 @@ public:
 
     bool isVector();
     bool isPoint();
-    int get(int index);
+    float get(int index);
     void print();
 };
